Move code into Cobra CLI structure.
